Module Poirot

Poirot is a grammar-based injection fuzzer for black box systems

author

Pierre-François Gimenez

Type

type grammar

The grammar manipulated by Poirot

type element

A symbol of the grammar, either a terminal or a nonterminal

type oracle_status

A variant type describing the oracle status

Main functions

val search : ?⁠verbose:bool -> ?⁠subst:(element, string) Stdlib.Hashtbl.t option -> ?⁠max_depth:int -> ?⁠forbidden_chars:char list -> ?⁠sgraph_fname:string option -> ?⁠qgraph_fname:string option -> (string option -> oracle_status) -> grammar -> element -> element list -> grammar option

search oracle g goal start returns the grammar of injection from the base grammar g according to the oracle oracle, starting from the trivial injection start.

parameter oracle

an oracle built with make_oracle_from_script or make_oracle_from_fun

parameter g

the grammar of the query language (e.g. SQL)

parameter goal

the goal of the search, i.e. the element (terminal or nonterminal) you seek to get in the grammar of injection. Poirot stops the search once it is reached.

parameter start

a element (terminal or nonterminal) that is a injection.

parameter verbose

(optional) makes Poirot verbose when true.

parameter subst

(optional) an Hashtable containing the semantics substitution.

parameter max_depth

(optional) modify the maximal depth of the search

parameter forbidden_chars

(optional) a list of forbidden characters. Such characters won't be used in injection. Useful to avoid escaped characters.

parameter sgraph_fname

(optional, for debug) export the search graph in graphviz dot format.

parameter qgraph_fname

(optional, for debug) export the quotient graph in graphviz dot format.

val quotient : grammar -> element list -> element list -> grammar

quotient g left_quotient right_quotient returns the grammar g after a left quotient by left_quotient and a right quotient by right_quotient. left_quotient and right_quotient can contain nonterminals or be empty.

val fuzzer : ?⁠subst:(element, string) Stdlib.Hashtbl.t option -> ?⁠complexity:int -> ?⁠goal:element option -> grammar -> string option

fuzzer ~complexity:10 ~goal:e g returns a word from g. If complexity is high, longer words should be generated. If complexity = 0, the returned word is deterministic (always the same) and should be short. If a goal is specified, then it will be included in the generated word. Beware: this is a primitive fuzzer.

Oracle functions

val make_oracle_from_script : ?⁠verbose:bool -> string -> string option -> oracle_status

make_oracle_from_script filename returns an oracle based on the script filename. The script must return error code 0 if the injection is lexically correct, 1 otherwise.

val make_oracle_from_fun : ?⁠verbose:bool -> (string -> int) -> string option -> oracle_status

make_oracle_from_fun f returns an oracle based on the function f. The function must return 0 if the injection is lexically correct, 1 otherwise.

Grammar manipulation functions

val to_uppercase : grammar -> grammar

to_uppercase g returns the grammar g with uppercased chars. The grammar is simplified, and nonterminals may disappear.

val to_lowercase : grammar -> grammar

to_lowercase g returns the grammar g with lowercased chars. The grammar is simplified, and nonterminals may disappear.

val set_axiom : grammar -> element -> grammar

set_axiom g new_axiom modifies the axiom of g with new_axiom.

val string_of_grammar : grammar -> string

string_of_grammar g returns the string representation of g.

I/O functions

val read_bnf_grammar : ?⁠unravel:bool -> string -> grammar

read_bnf_grammar filename reads a grammar from a bnf file filename.

val read_tokens : ?⁠unravel:bool -> string -> element list

read_tokens str reads a list of element from a string str.

val read_token : ?⁠unravel:bool -> string -> element

read_token str reads an element from a string str.

val export_antlr4 : string -> grammar -> unit

export_antlr4 filename g export the grammar g to antlr4 format into the file filename.g4

val read_subst : string -> (element, string) Stdlib.Hashtbl.t

read_subst filename read the semantics substitution from the file filename.