Module Poirot__Oracle

Type

type t

The type of an oracle.

type oracle_status =

| Syntax_error

| No_error

* syntax error: the injection is not syntactically correct * no error: the injection is syntactically correct

Oracle creation

val oracle_from_fun : float option -> (string -> oracle_status) -> t

make_oracle_from_fun f returns an oracle based on the function f. The function must return No_error if the injection is lexically correct, Syntax_error otherwise.

val oracle_from_script : float option -> float option -> string -> t

make_oracle_from_script filename returns an oracle based on the script filename. The script must return error code 0 if the injection is lexically correct, 180 otherwise.

val oracle_from_pf_sf : ?⁠oneline_comment:string option -> float option -> string -> string -> string -> t

make_oracle_from_pf_sf g_fname prefix suffix returns an oracle based on a given grammar quotiented by a prefix and a suffix. Useful to simulate a system.

Oracle call

val call : t -> string -> oracle_status

Call the oracle and handle the memoization, the interval, etc.

I/O

val save_mem : t -> string -> unit

Save the memoization into a file.

val load_mem : t -> string -> unit

Load the memoization into a file.

Other

val get_call_time : t -> float

val get_idle_time : t -> float

val get_call_nb : t -> int

val string_of_oracle_status : oracle_status -> string